Refactor datafusion utilities into a separate crate, rebase
