+++ rust/Cargo.toml
@@ -39,6 +39,7 @@ rust-version = "1.70"
 lance = { version = "=0.8.16", path = "./lance" }
 lance-arrow = { version = "=0.8.16", path = "./lance-arrow" }
 lance-core = { version = "=0.8.16", path = "./lance-core" }
+lance-datafusion = { version = "=0.8.16", path = "./lance-datafusion" }
 lance-datagen = { version = "=0.8.16", path = "./lance-datagen" }
 lance-index = { version = "=0.8.16", path = "./lance-index" }
 lance-linalg = { version = "=0.8.16", path = "./lance-linalg" }
