+++ rust/lance-datafusion/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "lance-datafusion"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+readme.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "Internal utilities used by other lance modules to simplify working with datafusion"
+
+[dependencies]
+arrow.workspace = true
+arrow-array.workspace = true
+arrow-schema.workspace = true
+datafusion.workspace = true
+datafusion-common.workspace = true
+datafusion-expr = "32.0.0"
+datafusion-physical-expr = { version = "32.0.0", default-features = false }
+futures.workspace = true
+lance-arrow.workspace = true
+lance-core.workspace = true
