+++ rust/lance-datafusion/src/expr.rs
@@ -12,30 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::{
-    cell::RefCell,
-    sync::{Arc, Mutex},
-};
+//! Utilities for working with datafusion expressions
 
-use arrow_array::RecordBatchReader;
-use arrow_schema::{DataType, Schema};
-use datafusion::{
-    execution::{
-        context::SessionState,
-        runtime_env::{RuntimeConfig, RuntimeEnv},
-    },
-    physical_plan::{
-        stream::RecordBatchStreamAdapter, DisplayAs, DisplayFormatType, ExecutionPlan,
-        SendableRecordBatchStream,
-    },
-    prelude::SessionConfig,
-};
-use datafusion_common::{DataFusionError, ScalarValue};
-use datafusion_physical_expr::Partitioning;
-
-use futures::{stream, TryStreamExt};
-use lance_arrow::SchemaExt;
-use lance_core::Result;
+use arrow_schema::DataType;
+use datafusion_common::ScalarValue;
 
 // This is slightly tedious but when we convert expressions from SQL strings to logical
 // datafusion expressions there is no type coercion that happens.  In other words "x = 7"
@@ -263,126 +243,4 @@ pub fn safe_coerce_scalar(value: &ScalarValue, ty: &DataType) -> Option<ScalarVa
         ScalarValue::Null => Some(value.clone()),
         _ => None,
     }
-}
-
-pub struct OneShotExec {
-    stream: Mutex<RefCell<Option<SendableRecordBatchStream>>>,
-    // We save off a copy of the schema to speed up formatting and so ExecutionPlan::schema & display_as
-    // can still function after exhuasted
-    schema: Arc<Schema>,
-}
-
-impl OneShotExec {
-    pub fn new(stream: SendableRecordBatchStream) -> Self {
-        let schema = stream.schema().clone();
-        Self {
-            stream: Mutex::new(RefCell::new(Some(stream))),
-            schema,
-        }
-    }
-}
-
-impl std::fmt::Debug for OneShotExec {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let val_guard = self.stream.lock().unwrap();
-        let stream = val_guard.borrow();
-        f.debug_struct("OneShotExec")
-            .field("exhausted", &stream.is_none())
-            .field("schema", self.schema.as_ref())
-            .finish()
-    }
-}
-
-impl DisplayAs for OneShotExec {
-    fn fmt_as(
-        &self,
-        t: datafusion::physical_plan::DisplayFormatType,
-        f: &mut std::fmt::Formatter,
-    ) -> std::fmt::Result {
-        let val_guard = self.stream.lock().unwrap();
-        let stream = val_guard.borrow();
-        match t {
-            DisplayFormatType::Default | DisplayFormatType::Verbose => {
-                let exhausted = if stream.is_some() { "" } else { "EXHUASTED" };
-                let columns = self
-                    .schema
-                    .field_names()
-                    .iter()
-                    .map(|s| s.to_string())
-                    .collect::<Vec<_>>();
-                write!(
-                    f,
-                    "OneShotStream: {} columns=[{}]",
-                    exhausted,
-                    columns.join(",")
-                )
-            }
-        }
-    }
-}
-
-/// Convert a sendable RecordBatchReader to a sendable RecordBatchStream
-pub fn reader_to_stream(reader: Box<dyn RecordBatchReader + Send>) -> SendableRecordBatchStream {
-    let schema = reader.schema().clone();
-    let stream = stream::iter(reader).map_err(DataFusionError::from);
-    Box::pin(RecordBatchStreamAdapter::new(schema, stream))
-}
-
-impl ExecutionPlan for OneShotExec {
-    fn as_any(&self) -> &dyn std::any::Any {
-        self
-    }
-
-    fn schema(&self) -> arrow_schema::SchemaRef {
-        self.schema.clone()
-    }
-
-    fn output_partitioning(&self) -> datafusion_physical_expr::Partitioning {
-        Partitioning::RoundRobinBatch(1)
-    }
-
-    fn output_ordering(&self) -> Option<&[datafusion_physical_expr::PhysicalSortExpr]> {
-        None
-    }
-
-    fn children(&self) -> Vec<Arc<dyn ExecutionPlan>> {
-        vec![]
-    }
-
-    fn with_new_children(
-        self: Arc<Self>,
-        _children: Vec<Arc<dyn ExecutionPlan>>,
-    ) -> datafusion_common::Result<Arc<dyn ExecutionPlan>> {
-        todo!()
-    }
-
-    fn execute(
-        &self,
-        _partition: usize,
-        _context: Arc<datafusion::execution::TaskContext>,
-    ) -> datafusion_common::Result<SendableRecordBatchStream> {
-        let mut val_guard = self.stream.lock().unwrap();
-        let stream = val_guard.get_mut();
-        let stream = stream.take();
-        if let Some(stream) = stream {
-            Ok(stream)
-        } else {
-            panic!("Attempt to use OneShotExec more than once");
-        }
-    }
-
-    fn statistics(&self) -> datafusion_common::Statistics {
-        todo!()
-    }
-}
-
-pub fn execute_plan(plan: Arc<dyn ExecutionPlan>) -> Result<SendableRecordBatchStream> {
-    let session_config = SessionConfig::new();
-    let runtime_config = RuntimeConfig::new();
-    let runtime_env = Arc::new(RuntimeEnv::new(runtime_config)?);
-    let session_state = SessionState::new_with_config_rt(session_config, runtime_env);
-    // NOTE: we are only executing the first partition here. Therefore, if
-    // the plan has more than one partition, we will be missing data.
-    assert_eq!(plan.output_partitioning().partition_count(), 1);
-    Ok(plan.execute(0, session_state.task_ctx())?)
 }
