+++ rust/lance-datafusion/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod chunker;
+pub mod exec;
+pub mod expr;
