+++ rust/lance/src/dataset/fragment.rs
@@ -33,7 +33,7 @@ use lance_core::{
     },
     Error, Result, ROW_ID,
 };
-use lance_index::util::chunker::chunk_stream;
+use lance_datafusion::chunker::chunk_stream;
 use object_store::path::Path;
 use snafu::{location, Location};
 use uuid::Uuid;
